Provide a way to not rebuild `roxctl` when running tests
to speed up local iteration.
Not intended for CI and not documented because it's only for curious
minds ;-P

+++ tests/roxctl/bats-tests/local/roxctl-helm-output-secured-cluster-services-development.bats
@@ -6,7 +6,7 @@ out_dir=""
 
 setup_file() {
   # remove binaries from the previous runs
-  rm -f "${tmp_roxctl}"/roxctl*
+  [[ -n "$NO_BATS_ROXCTL_REBUILD" ]] || rm -f "${tmp_roxctl}"/roxctl*
 
   echo "Testing roxctl version: '$(roxctl-development version)'" >&3
   command -v yq || skip "Tests in this file require yq"
